+++ tests/TestCase.php
@@ -20,4 +20,20 @@ abstract class TestCase extends OrchestraTestCase
             DependenciesServiceProvider::class,
         ];
     }
+
+    /**
+     * Retrieve an array of packages.
+     *
+     * @return array
+     */
+    public function packageProvider(): array
+    {
+        return [
+            ['sfneal/actions'],
+            ['sfneal/aws-s3-helpers'],
+            ['sfneal/laravel-helpers'],
+            ['laravel/framework'],
+            ['spatie/laravel-view-models'],
+        ];
+    }
 }