+++ CHANGELOG.md
@@ -7,3 +7,8 @@ All notable changes to `dependencies` will be documented in this file
 - initial pre-release
 - make config file, service provider, service & repository
 - add dependencies (laravel & sfneal/string-helpers)
+
+
+## 0.1.1 - 2021-07-08
+- add url & response contains assertions
+- add packages to the dataProvider